add pagination

+++ frontend/hooks/usePosts.tsx
@@ -1,13 +1,23 @@
-import useSWR, { mutate } from "swr";
 import { api } from "../utils/api";
 import { Post } from "@/types/Post";
 import useAuth from "./useAuth";
+import useSWRInfinite from "swr/infinite";
+import { useMemo } from "react";
 
 const usePosts = (id?: number) => {
+  const PAGE_SIZE = 6;
+
   const { user, loggedIn } = useAuth();
 
-  const { data, error } = useSWR<Post[], Error>(
-    "/posts/",
+  const {
+    data: pages,
+    error,
+    isLoading,
+    size,
+    setSize,
+    mutate,
+  } = useSWRInfinite(
+    (index) => `/posts/?per_page=${PAGE_SIZE}&page=${index + 1}`,
     async (url: string) => {
       const response = await api.get(url);
 
@@ -15,37 +25,61 @@ const usePosts = (id?: number) => {
     }
   );
 
+  const posts: Post[] = useMemo(() => {
+    return pages ? [].concat(...pages) : [];
+  }, [pages]);
+
+  const isReachingEnd = useMemo(() => {
+    return pages && pages[pages.length - 1]?.length < PAGE_SIZE;
+  }, [pages]);
+
   const deletePost = () => {
-    if (id) {
+    if (id && !isLoading) {
       api.delete(`/posts/${id}`).then(() => {
-        mutate("/posts/");
+        mutate(
+          pages?.map((page) => page.filter((post: Post) => post.id !== id))
+        );
       });
     }
   };
 
-  const createPost = (data: Partial<Post>) => {
-    if (loggedIn) {
-      api.post("/posts/", { ...data, username: user?.username }).then(() => {
-        mutate("/posts/");
+  const createPost = (newData: Partial<Post>) => {
+    if (loggedIn && !isLoading) {
+      api.post("/posts/", { ...newData, username: user?.username }).then(() => {
+        if (pages?.length) {
+          mutate([...pages, [newData]]);
+        } else {
+          mutate([[newData]]);
+        }
       });
     }
   };
 
-  const updatePost = (data: Partial<Post>) => {
-    if (id) {
-      api.patch(`/posts/${id}`, data).then(() => {
-        mutate("/posts/");
+  const updatePost = (newData: Partial<Post>) => {
+    if (id && !isLoading) {
+      api.patch(`/posts/${id}`, newData).then(() => {
+        mutate(
+          pages?.map((page) =>
+            page.map((post: Post) => (post.id === id ? newData : post))
+          )
+        );
       });
     }
+  };
+
+  const loadMore = () => {
+    setSize(size + 1);
   };
 
   return {
-    posts: data,
-    isLoading: !error && !data,
+    posts,
+    isLoading,
     isError: error,
     deletePost,
     createPost,
     updatePost,
+    loadMore,
+    isReachingEnd,
   };
 };
